pyproject.toml
[tool.poetry]
name = "bobot"
version = "0.1.0"
description = "Boot para discord"
authors = ["Anderson Rodrigues <eusouanderson@outlook.com>"]
readme = "README.md"
packages = [{ include = "bobot", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
aiosignal = "1.3.1"
async-timeout = "4.0.2"
attrs = "23.1.0"
cachetools = "5.3.1"
certifi = "2023.5.7"
charset-normalizer = "3.1.0"
click = "8.1.3"
cycler = "0.11.0"
discord-py = "2.3.0"
fonttools = "4.40.0"
google-api-core = "2.11.1"
google-api-python-client = "2.90.0"
google-auth = "2.20.0"
google-auth-httplib2 = "0.1.0"
googleapis-common-protos = "1.59.1"
httplib2 = "0.22.0"
idna = "3.4"
mccabe = "0.6.1"
mypy-extensions = "1.0.0"
packaging = "23.1"
pathspec = "0.11.1"
pillow = "9.5.0"
platformdirs = "3.7.0"
protobuf = "4.23.3"
pyasn1 = "0.5.0"
pyasn1-modules = "0.3.0"
pycodestyle = "2.8.0"
pyflakes = "2.4.0"
pyparsing = "3.1.0"
python-dateutil = "2.8.2"
requests = "2.31.0"
rsa = "4.9"
six = "1.16.0"
tomli = "2.0.1"
tqdm = "4.65.0"
uritemplate = "4.1.1"
urllib3 = "1.26.16"
pyarmor = "^8.5.8"
aiohttp = "^3.9.5"
python-dotenv = "^1.0.1"
httpx = "^0.27.2"

[tool.poetry.group.dev.dependencies]
black = "22.1.0"
blue = "0.9.1"
flake8 = "4.0.1"
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
pytest-cov = "^7.0.0"

[tool.poetry.scripts]
bobot = "bobot.__main__:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
